Gunnar: minor edits
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1078 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/FermionOnSphereWithSpin.cc
@@ -107,8 +107,8 @@ FermionOnSphereWithSpin::FermionOnSphereWithSpin (int nbrFermions, int totalLz, 
   this->GenerateLookUpTable(memory);
   
 #ifdef __DEBUG__
-  int UsedMemory = 0;
-  UsedMemory += this->HilbertSpaceDimension * (sizeof(unsigned long) + sizeof(int));
+  long UsedMemory = 0;
+  UsedMemory += (long) this->HilbertSpaceDimension * (sizeof(unsigned long) + sizeof(int));
   cout << "memory requested for Hilbert space = ";
   if (UsedMemory >= 1024)
     if (UsedMemory >= 1048576)
